client-toolkit/screencopy: Split dispatch impls into a file
